TestDNSLogging: rework to prevent CI flakes
Previously, the test initiated a pod to issue 'dig' queries and
fetched the list of coredns pods. For each coredns pod, a 'dig'
request was made, and logs for the "dns" container were retrieved to
assert the functionality of logging.

However, queries from any pod prefer the coredns pod on the same node,
leading to potential failures when the first checked coredns pod
wasn't on the same node as the test pod.

The test has been reworked to remove the test pod and execute the
following command for each coredns pod:

  oc exec <coredns-pod> -- dig @localhost -p 5353 <hostname>

Existing test assertions remain unchanged; the "dns" container logs
are retrieved, and the test passes if a logged entry for <hostname> is
found.