OCPBUGS-15605: Update bufsize to 1232 bytes
As endorsed at [DNS Flag Day](http://www.dnsflagday.net/2020/),
the DNS Community recommends a bufsize setting of 1232 as a
safe default that supports larger payloads, while generally
avoiding IP fragmentation on most networks. This is particularly
relevant for payloads like those generated by DNSSEC, which tend to
be larger.

Previously, CoreDNS always used the EDNS0 extension, which enables
UDP-based DNS queries to exceed 512 bytes, when CoreDNS forwarded
DNS queries to an upstream name server, and so OpenShift specified
a bufsize setting of 512 to maintain compatibility with applications
and name servers that did not support the EDNS0 extension.

Currently, CoreDNS no longer uses EDNS0 when forwarding requests
if the original client request did not use EDNS0, and so the
reasoning for using a bufsize setting of 512 no longer applies.
By increasing the bufsize setting to the recommended value of 1232 bytes,
we can enhance DNS performance by decreasing the probability of DNS
truncations.

This PR updates the bufsize setting for CoreDNS from 512 to 1232 bytes.