pkg/operator/controller: Only bump lastTransitionTime on 'status' changes
From the API docs [1]:

  lastTransitionTime is the time of the last update to the current status property.

And the library-go SetStatusCondition implementation includes [2]:

  if existingCondition.Status != newCondition.Status {
    existingCondition.Status = newCondition.Status
    existingCondition.LastTransitionTime = metav1.NewTime(time.Now())
  }

  existingCondition.Reason = newCondition.Reason
  existingCondition.Message = newCondition.Message

The motivation for that behavior is that it's often more useful to
know "how long has this resource been Degraded=True?" and similar than
it is to know how long it has had exactly the same
status/reason/message.  Messages in particular can be fairly mutable
("# of # pods available", etc.), and changes there do not necessarily
represent fundamental shifts in the underlying issue.

[1]: https://github.com/openshift/api/blob/81f778f3b3ec31c1dd344e795620a8fbaf2d9a51/config/v1/types_cluster_operator.go#L129
[2]: https://github.com/openshift/library-go/blob/c515269de16e5e239bd6e93e1f9821a976bb460b/pkg/config/clusteroperator/v1helpers/status.go#L29C1-L35C50