pkg/operator/controller/dns_status: Allow 10% unavailable before going Degraded=True
5c4f0ad3430a (Set DNS DaemonSet's maxSurge value to 10%, 2023-03-23, #358)
pivoted the dns-default DaemonSet from 10% maxUnavailable to 10%
maxSurge.  That left this "is this DaemonSet sad?" checker defaulting
to "if there are any more than one unready DNS pods, we're sad".  But
in many clusters, especially large ones with the DaemonSet trying to
schedule pods on many compute nodes, there are often more than 1
unready pod, and going Degraded=True is distracting noise, and not
something that has a quality-of-service impact.  This commit adjusts
the default to the old 10% maxUnavailable, so we have a less twitchy
value we can use for surging DaemonSets.  We can't set that on the
DaemonSet itself, because as described in 5c4f0ad3430a, API validation
prevents us from setting maxSurge and maxUnavailable simultaneously.