Add RBAC for reading featuregates
Commit d32ca6bbe050c0732b5a6f3ecb10e50bf2da8a25 in the release-4.16
branch added RBAC for featuregates in the cluster role for the DNS
operator as well as a configInformers.Config().V1().FeatureGates() call
in the DNS operator in OpenShift 4.16.  This causes issues during
upgrade of HyperShift clusters from 4.15 to 4.16 as sometimes the
deployment of the DNS operator may get upgraded before the cluster role,
which causes the DNS operator to crashloop (by design) temporarily
during the upgrade.

This commit adds the required RBAC to the cluster role in 4.15 so that
the DNS operator doesn't crashloop during the upgrade of HyperShift
clusters from 4.15 to 4.16.

This commit fixes OCPBUGS-32093.

https://issues.redhat.com//browse/OCPBUGS-32093